+++ packages/cli/webpack.config.js
@@ -1,6 +1,5 @@
 const path = require('path');
 const webpack = require('webpack');
-const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 const CopyPlugin = require('copy-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const { ModuleFederationPlugin } = require('webpack').container;
@@ -61,7 +60,7 @@ module.exports = [
       filename: 'index.js',
       path: path.resolve(__dirname, 'dist/shell'),
     },
-    mode: 'development',
+    mode: 'production',
     module: {
       rules: [
         {